Move errors to own file. Move monitor to own file.
